tools/company-graph/requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
pyyaml>=6.0.1

# Japanese NLP (optional – enables GiNZA NER for better company name extraction)
# Install with: pip install spacy ja-ginza
# spacy>=3.7.0
# ja-ginza>=5.1.0

# Development / testing
pytest>=8.0.0
pytest-cov>=5.0.0
